@@ -3,8 +3,8 @@ description: Demonstrates how to use the url_launcher plugin.
 publish_to: none
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: '>=2.18.0 <3.0.0'
+  flutter: ">=3.3.0"
 
 dependencies:
   flutter:
