@@ -2,11 +2,11 @@ name: url_launcher_ios
 description: iOS implementation of the url_launcher plugin.
 repository: https://github.com/flutter/plugins/tree/main/packages/url_launcher/url_launcher_ios
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+url_launcher%22
-version: 6.0.18
+version: 6.1.0
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: '>=2.18.0 <3.0.0'
+  flutter: ">=3.3.0"
 
 flutter:
   plugin:
